temporarily check in of 'revert z-buffer from ints back to shorts'

+++ Jmol/src/org/jmol/g3d/Graphics3D.java
@@ -70,7 +70,7 @@ final public class Graphics3D {
   int slab, depth;
   int xLast, yLast;
   int[] pbuf;
-  int[] zbuf;
+  short[] zbuf;
 
   int clipX;
   int clipY;
@@ -85,7 +85,7 @@ final public class Graphics3D {
 
   Font3D font3dCurrent;
 
-  final static int ZBUFFER_BACKGROUND = Platform3D.ZBUFFER_BACKGROUND;
+  final static short ZBUFFER_BACKGROUND = Platform3D.ZBUFFER_BACKGROUND;
 
   /**
    * Allocates a g3d object
@@ -1162,7 +1162,7 @@ final public class Graphics3D {
       return;
     int offset = y * width + x;
     if (z < zbuf[offset]) {
-      zbuf[offset] = z;
+      zbuf[offset] = (short) z;
       pbuf[offset] = argbCurrent;
     }
   }
@@ -1173,7 +1173,7 @@ final public class Graphics3D {
     int z = screen.z; if (z < slab || z > depth) return;
     int offset = y * width + x;
     if (z < zbuf[offset]) {
-      zbuf[offset] = z;
+      zbuf[offset] = (short) z;
       pbuf[offset] = argbCurrent;
     }
   }
@@ -1183,7 +1183,7 @@ final public class Graphics3D {
       return;
     int offset = y * width + x;
     if (z < zbuf[offset]) {
-      zbuf[offset] = z;
+      zbuf[offset] = (short) z;
       pbuf[offset] = argb;
     }
   }
@@ -1193,7 +1193,7 @@ final public class Graphics3D {
       return;
     int offset = y * width + x;
     if (z < zbuf[offset]) {
-      zbuf[offset] = z;
+      zbuf[offset] = (short) z;
       pbuf[offset] = argb;
     }
   }
@@ -1205,7 +1205,7 @@ final public class Graphics3D {
       return;
     int offset = y * width + x;
     if (z < zbuf[offset]) {
-      zbuf[offset] = z;
+      zbuf[offset] = (short) z;
       pbuf[offset] = argb;
     }
   }
@@ -1215,7 +1215,7 @@ final public class Graphics3D {
       return;
     int offset = y * width + x;
     if (z < zbuf[offset]) {
-      zbuf[offset] = z;
+      zbuf[offset] = (short) z;
       pbuf[offset] = getColixArgb(colix);
     }
   }
@@ -1223,7 +1223,7 @@ final public class Graphics3D {
   void plotPixelUnclipped(int x, int y, int z) {
     int offset = y * width + x;
     if (z < zbuf[offset]) {
-      zbuf[offset] = z;
+      zbuf[offset] = (short) z;
       pbuf[offset] = argbCurrent;
     }
   }
@@ -1231,7 +1231,7 @@ final public class Graphics3D {
   void plotPixelUnclipped(int argb, int x, int y, int z) {
     int offset = y * width + x;
     if (z < zbuf[offset]) {
-      zbuf[offset] = z;
+      zbuf[offset] = (short) z;
       pbuf[offset] = argb;
     }
   }
@@ -1257,7 +1257,7 @@ final public class Graphics3D {
     }
     while (offsetPbuf < offsetMax) {
       if (z < zbuf[offsetPbuf]) {
-        zbuf[offsetPbuf] = z;
+        zbuf[offsetPbuf] = (short) z;
         pbuf[offsetPbuf] = argbCurrent;
       }
       offsetPbuf += step;
@@ -1302,7 +1302,7 @@ final public class Graphics3D {
         if (! isTranslucent || (flipflop = !flipflop)) {
           int z = zScaled >> 10;
           if (z >= slab && z <= depth && z < zbuf[offsetPbuf]) {
-            zbuf[offsetPbuf] = z;
+            zbuf[offsetPbuf] = (short) z;
             seed = ((seed << 16) + (seed << 1) + seed) & 0x7FFFFFFF;
             int bits = (seed >> 16) & 0x07;
             pbuf[offsetPbuf] = (bits == 0
@@ -1324,7 +1324,7 @@ final public class Graphics3D {
         if (! isTranslucent || (flipflop = !flipflop)) {
           int z = zScaled >> 10;
           if (z >= slab && z <= depth && z < zbuf[offsetPbuf]) {
-            zbuf[offsetPbuf] = z;
+            zbuf[offsetPbuf] = (short) z;
             pbuf[offsetPbuf] = (0xFF000000 |
                                 (rScaled & 0xFF0000) |
                                 (gScaled & 0xFF00) |
@@ -1362,7 +1362,7 @@ final public class Graphics3D {
         while (--count >= 0) {
           int z = zScaled >> 10;
           if (z < zbuf[offsetPbuf]) {
-            zbuf[offsetPbuf] = z;
+            zbuf[offsetPbuf] = (short) z;
             seed = ((seed << 16) + (seed << 1) + seed) & 0x7FFFFFFF;
             int bits = (seed >> 16) & 0x07;
             pbuf[offsetPbuf] = (bits == 0
@@ -1379,7 +1379,7 @@ final public class Graphics3D {
           if (flipflop) {
             int z = zScaled >> 10;
             if (z < zbuf[offsetPbuf]) {
-              zbuf[offsetPbuf] = z;
+              zbuf[offsetPbuf] = (short) z;
               seed = ((seed << 16) + (seed << 1) + seed) & 0x7FFFFFFF;
               int bits = (seed >> 16) & 0x07;
               pbuf[offsetPbuf] = (bits == 0
@@ -1443,7 +1443,7 @@ final public class Graphics3D {
         if (! isTranslucent || (flipflop = !flipflop)) {
           int z = zScaled >> 10;
           if (z < zbuf[offsetPbuf]) {
-            zbuf[offsetPbuf] = z;
+            zbuf[offsetPbuf] = (short) z;
             pbuf[offsetPbuf] = (0xFF000000 |
                                 (rScaled & 0xFF0000) |
                                 (gScaled & 0xFF00) |
@@ -1464,7 +1464,7 @@ final public class Graphics3D {
     if (! isTranslucent) {
       while (--count >= 0) {
         if (z < zbuf[offsetPbuf]) {
-          zbuf[offsetPbuf] = z;
+          zbuf[offsetPbuf] = (short) z;
           pbuf[offsetPbuf] = argbCurrent;
         }
         ++offsetPbuf;
@@ -1476,7 +1476,7 @@ final public class Graphics3D {
           return;
       do {
         if (z < zbuf[offsetPbuf]) {
-          zbuf[offsetPbuf] = z;
+          zbuf[offsetPbuf] = (short) z;
           pbuf[offsetPbuf] = argbCurrent;
         }
         offsetPbuf += 2;
@@ -1503,7 +1503,7 @@ final public class Graphics3D {
       int alpha = pixel & 0xFF000000;
       if (alpha >= 0x80000000) {
         if (z < zbuf[offsetPbuf]) {
-          zbuf[offsetPbuf] = z;
+          zbuf[offsetPbuf] = (short) z;
           pbuf[offsetPbuf] = pixel;
           }
       }
@@ -1519,7 +1519,7 @@ final public class Graphics3D {
       int alpha = pixel & 0xFF000000;
       if ((alpha & 0x80000000) != 0) {
         if (z < zbuf[offsetPbuf]) {
-          zbuf[offsetPbuf] = z;
+          zbuf[offsetPbuf] = (short) z;
           pbuf[offsetPbuf] = pixel;
         }
       }
@@ -1589,7 +1589,7 @@ final public class Graphics3D {
         continue;
       int offset = y * width + x;
       if (z < zbuf[offset]) {
-        zbuf[offset] = z;
+        zbuf[offset] = (short) z;
         pbuf[offset] = argb;
       }
     }
@@ -1606,7 +1606,7 @@ final public class Graphics3D {
         continue;
       int offset = y * width + x;
       if (z < zbuf[offset]) {
-        zbuf[offset] = z;
+        zbuf[offset] = (short) z;
         //        pbuf[offset] = getColixArgb(colix);
         pbuf[offset] = shades[intensities[j]];
       }
